requirements.txt
pyrofork==2.3.45
tgcrypto==1.2.5
motor>=3.7.1
Flask==1.1.2
gunicorn==20.1.0
Jinja2==3.0.3
Werkzeug==2.0.2
itsdangerous==2.0.1
python-dotenv>=1.1.0
# Agar Mongo URI mongodb+srv:// use karte ho
dnspython>=2.4
